+ Wallet system and UIManager
In this commit, the Wallet script is introduced to handle payments and track the money amount.
The UIManager script is added to manage UI changes.
The FinishSession function is implemented in FruitsController to handle payment processing at the end of a session.
Additionally, credit and bet texts have been added to the scene to display relevant information.

+++ Assets/Scripts/FruitSettings.cs
@@ -8,7 +8,7 @@ public class FruitSettings : ScriptableObject
     public Sprite sprite;
     public float spawnRate;
     public bool bonus;
-    public float eightNinePaymentMultiplier;
-    public float tenElevenPaymentMultiplier;
-    public float twelveAndAbovePaymentMultiplier;
+    public double eightNinePaymentMultiplier;
+    public double tenElevenPaymentMultiplier;
+    public double twelveAndAbovePaymentMultiplier;
 }
